requirements.txt
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
flake8==3.9.2
Flask==2.0.1
future==0.18.2
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
line-bot-sdk==1.19.0
MarkupSafe==2.0.1
mccabe==0.6.1
mysql-connector-python==8.2.0
protobuf==4.21.12
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
requests==2.25.1
toml==0.10.2
urllib3==1.26.5
Werkzeug==2.0.1
